move build output json test
